Creates controller designs, edits some routes & designs controller
This commit creates the controller Designs 
The generation was incomplete so I bring changes to the files.
The code inside the controller makes sure that only the admin can edit, create or delete a design and such things. It also hides some buttons to standard users. 
The routes are edited to add designs routes.